SSR warn for no VUE_ENV and add tests
- Added test coverage for data observation and VUE environment variable
- Moved Test helpers into own folder for reusability
