+++ src/server/create-renderer.js
@@ -1,11 +1,19 @@
 import RenderStream from './render-stream'
 import { createRenderFunction } from './render'
+import { warn } from 'core/util/debug'
 
 export function createRenderer ({
   modules = [],
   directives = {},
   isUnaryTag = (() => false)
 } = {}) {
+  if (process.env.VUE_ENV !== 'server') {
+    warn(
+      'You are using createRenderer without setting VUE_ENV enviroment variable to "server". ' +
+      'It is recommended to set VUE_ENV=server this will help rendering performance, ' +
+      'by turning data observation off.'
+    )
+  }
   const render = createRenderFunction(modules, directives, isUnaryTag)
   return {
     renderToString (component) {
